posixtm tests: Fix warnings seen on MidnightBSD/x86.
* tests/test-posixtm.c (main): Convert two 'time_t' values to 'long' for
printing.

+++ tests/test-posixtm.c
@@ -178,7 +178,7 @@ main (void)
       if (t_out != t_exp)
         {
           printf ("%s mismatch (-: actual; +:expected)\n-%12ld\n+%12ld\n",
-                  T[i].in, t_out, t_exp);
+                  T[i].in, (long) t_out, (long) t_exp);
           fail = 1;
         }
     }
